delete uncached comments using partials

+++ src/bot/events/message/message.ts
@@ -1,8 +1,8 @@
-import { Event } from '../eventHandler.js';
+import { Event } from '../../eventHandler.js';
 import { Message, TextChannel, MessageEmbed } from 'discord.js';
-import { commands, settings } from '../bot.js';
-import { runCommand } from '../commandHandler.js';
-import Ticket from '../../db/models/Ticket.js';
+import { commands, settings } from '../../bot.js';
+import { runCommand } from '../../commandHandler.js';
+import Ticket from '../../../db/models/Ticket.js';
 import mongoose from 'mongoose';
 
 export const event: Event = {
@@ -22,29 +22,32 @@ export const event: Event = {
 
 async function ticketComment(message: Message) {
     const { channel, member, content } = message;
-    if (!(channel instanceof TextChannel) || !channel.topic || !member) return;
+    if (message.partial || !(channel instanceof TextChannel) || !channel.topic || !member) return;
 
     const id = channel.topic.split(' | ')[0];
 
     const comment = {
         _id: new mongoose.Types.ObjectId(),
         author: member.id,
+        message: '',
         content,
         timestamp: new Date().toISOString(),
     };
 
-    await Ticket.findByIdAndUpdate(id, {
-        $push: {
-            comments: comment,
-        },
-    });
-
     await message.delete();
-    channel.send(
+    const commentMessage = await channel.send(
         new MessageEmbed()
             .setAuthor(member.user.username + '#' + member.user.discriminator, member.user.avatarURL() || undefined)
             .setFooter('ID: ' + comment._id)
             .setTimestamp(new Date(comment.timestamp))
             .setDescription(content)
     );
+
+    comment.message = commentMessage.id;
+
+    await Ticket.findByIdAndUpdate(id, {
+        $push: {
+            comments: comment,
+        },
+    });
 }
